+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Switch Telehealth/Online depending on provider [#3241](https://github.com/rokwire/illinois-app/issues/3241).
 - Show person's name in Appointments card, clean up card's layout [#3241](https://github.com/rokwire/illinois-app/issues/3241).
+- Do not show safer McKinley link in Detail panel for other providers [#3244](https://github.com/rokwire/illinois-app/issues/3244).
+- Updated number of advisors text [#3246](https://github.com/rokwire/illinois-app/issues/3246).
+- Rename "time slot" to "available appointment" [#3248](https://github.com/rokwire/illinois-app/issues/3248).
 
 ## [4.3.55] - 2023-05-04
 ### Changed